Revert "Protect min_heap_push_ against integer overflow."
This patch breaks the ABI compatibility, due to min_heap_idx expansion
[1]

And since major distros did not updated to 2.1.10 yet, this patch will
be reverted:
- debian:     https://packages.debian.org/search?keywords=libevent-dev
- ubuntu:     https://packages.ubuntu.com/search?keywords=libevent-dev
- fedora:     https://apps.fedoraproject.org/packages/libevent-devel

Also there is one that upgraded already:
- archlinux:  https://www.archlinux.org/packages/?q=libevent
But archlinux is for developers, so it should be fine I guess.

  [1]: https://abi-laboratory.pro/index.php?view=objects_report&l=libevent&v1=2.1.9&v2=2.1.10

    - struct event_base
      Change: Size of this type has been changed from 672 bytes to 664 bytes.
      Effect: The fields or parameters of such data type may be incorrectly initialized or accessed by old client applications.

abi-compliance-checker diff with 2.1.9 before this patch:
  Binary compatibility: 69.2%
  Source compatibility: 100%
  Total binary compatibility problems: 1, warnings: 1
  Total source compatibility problems: 0, warnings: 0

after:
  Binary compatibility: 100%
  Source compatibility: 100%
  Total binary compatibility problems: 0, warnings: 0
  Total source compatibility problems: 0, warnings: 0

This reverts commit 0b46bb8cc9c0337b5fa0186d9cb031ff4f4ceb9a

+++ include/event2/event_struct.h
@@ -126,7 +126,7 @@ struct event {
 	/* for managing timeouts */
 	union {
 		TAILQ_ENTRY(event) ev_next_with_common_timeout;
-		size_t min_heap_idx;
+		int min_heap_idx;
 	} ev_timeout_pos;
 	evutil_socket_t ev_fd;
